+++ UnitTests/BraintreeShopperInsightsTests/GenerateCustomerRecommendationsGraphQLBody_Tests.swift
@@ -19,6 +19,10 @@ class GenerateCustomerRecommendationsGraphQLBody_Tests: XCTestCase {
                 amount: "12.00",
                 currencyCode: "USD"
             )
+        ],
+        payPalCampaigns: [
+            ShopperInsightsCampaign(id: "campaign-123-id"),
+            ShopperInsightsCampaign(id: "campaign-456-id")
         ]
     )
     let expectedQuery = """
@@ -46,12 +50,16 @@ class GenerateCustomerRecommendationsGraphQLBody_Tests: XCTestCase {
         let input = variables?["input"] as? [String: Any]
         let customer = input?["customer"] as? [String: Any]
         let purchaseUnits = input?["purchaseUnits"] as? [[String: Any]]
+        let payPalCampaigns = input?["paypalCampaigns"] as? [[String: Any]]
         let amount = purchaseUnits?.first?["amount"] as? [String: Any]
         
         XCTAssertEqual(jsonObject["query"] as? String, expectedQuery)
         XCTAssertEqual(input?["sessionId"] as? String, sessionID)
         XCTAssertEqual(customer?["hashedEmail"] as? String, "test-hashed-email.com")
         XCTAssertEqual(customer?["paypalAppInstalled"] as? Bool, true)
+        XCTAssertEqual(payPalCampaigns?.count, 2)
+        XCTAssertEqual(payPalCampaigns?[0]["id"] as? String, "campaign-123-id")
+        XCTAssertEqual(payPalCampaigns?[1]["id"] as? String, "campaign-456-id")
         XCTAssertEqual(amount?["value"] as? String, "5.00")
     }
     
@@ -61,7 +69,8 @@ class GenerateCustomerRecommendationsGraphQLBody_Tests: XCTestCase {
             hashedPhoneNumber: nil,
             payPalAppInstalled: nil,
             venmoAppInstalled: nil,
-            purchaseUnits: nil
+            purchaseUnits: nil,
+            payPalCampaigns: []
         )
         
         let body = GenerateCustomerRecommendationsGraphQLBody(request: request, sessionID: sessionID)
@@ -74,9 +83,11 @@ class GenerateCustomerRecommendationsGraphQLBody_Tests: XCTestCase {
         let input = variables?["input"] as? [String: Any]
         let customer = input?["customer"] as? [String: Any]
         let purchaseUnits = input?["purchaseUnits"] as? [[String: Any]]
+        let payPalCampaigns = input?["paypalCampaigns"] as? [[String: Any]]
         
         XCTAssertNotNil(customer)
         XCTAssertNil(purchaseUnits)
+        XCTAssertNil(payPalCampaigns)
         XCTAssertEqual(input?["sessionId"] as? String, sessionID)
         XCTAssertEqual(jsonObject["query"] as? String, expectedQuery)
     }